Major Overhaul [See Description].
The program will now run even if hardware is not connected. (This does not protect from hardware that disconnects while the program is running). The motors are considered a direct extension of the Arduino, meaning that if the arduino is connected, all 4 motors are presumed to be connected. On a similar note. I also removed the Hardware Module becuase it was no longered required. Finally I moved the changeInterval() Function to the ROV_functions module. (Other minor formatting and changes were made).