Gamepad can now be Reconnected!
As long as we continue to use this model gamepad, we are able to reconnect the controller at anytime. The Motor Joystick takes priority over the arm controller (by design). If the gamepad disconnects, then then it will not attempt to reconnect until the joystick is disconnected (We don't have to physically disconnect the joystick. We just have to choose a button to press to mark the controller as disconnected) I did this purposely so that we don't loose control of the motors, if the gamepad fails. While we should stick with the PS3 controller; this code should be editable for any controller with motion tracking.