Added Servo Support.
NOTE: WITHOUT THE HARDWARE THIS CODE IS NOT TESTED. SERVO ROTATION MAY BE REVERSED UNTIL THEN. That being said, the code should work, It will just have to be adjusted once servos are connected. On one of the previous commits, I mentioned that we would probably need 2 arduino boards. I no longer believe this to be the case. Expect another commit in a few moments, where I'll be renaming motorArduino back to arduino.