Added Screenshot Functionality.
Added functionality to take screenshots of the camera, and to display them on the screen. The code to write the screenshots to the window has been commented out becuase I have not adjusted the screenshot size, nor have I found the proper coordinates for where they will be on the window.